fix: blank lines due to common checklist items
