+++ fbw-a380x/src/systems/instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata24.ts
@@ -1,11 +1,8 @@
 ﻿// Copyright (c) 2024 FlyByWire Simulations
 // SPDX-License-Identifier: GPL-3.0
 
-import {
-  AbnormalProcedure,
-  ChecklistLineStyle,
-  FMS_PRED_UNRELIABLE_CHECKLIST_ITEM,
-} from 'instruments/src/MsfsAvionicsCommon/EcamMessages';
+import { AbnormalProcedure, ChecklistLineStyle } from 'instruments/src/MsfsAvionicsCommon/EcamMessages';
+import { FMS_PRED_UNRELIABLE_CHECKLIST_ITEM } from '../CommonChecklistItems';
 
 // Convention for IDs:
 // First two digits: ATA chapter
