+++ fbw-a380x/src/systems/instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata31-32-33.ts
@@ -6,9 +6,8 @@ import {
   ChecklistLineStyle,
   DeferredProcedure,
   DeferredProcedureType,
-  FMS_PRED_UNRELIABLE_CHECKLIST_ITEM,
-  STEER_ENDURANCE_LIMITED_CHECKLIST_ITEM,
 } from 'instruments/src/MsfsAvionicsCommon/EcamMessages';
+import { FMS_PRED_UNRELIABLE_CHECKLIST_ITEM, STEER_ENDURANCE_LIMITED_CHECKLIST_ITEM } from '../CommonChecklistItems';
 
 // Convention for IDs:
 // First two digits: ATA chapter
